# back/public/api-docs.yaml
openapi: 3.0.0
info:
  title: 'API ProjetPro'
  description: "API REST pour l'application ProjetPro avec système d'authentification sécurisé.\n\n## Fonctionnalités principales:\n- Authentification avec protection anti-force brute\n- Gestion des tokens CSRF\n- Upload d'images vers Azure Blob Storage\n- Logging automatique des tentatives de connexion\n- Système de blocage par IP et par login\n\n## Sécurité:\n- Protection CSRF obligatoire\n- Tokens d'authentification JWT\n- Blocage automatique après échecs répétés\n\n## Headers requis:\n- `X-CSRF-Token`: Token CSRF pour les requêtes POST\n- `Authorization: Bearer <token>`: Token JWT pour l'authentification\n- `Content-Type: application/json`\n"
  contact:
    name: 'Support API'
    email: support@projetpro.com
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'Serveur de développement'
  -
    url: /api
    description: 'API Production'
paths:
  /api/csrfToken:
    get:
      tags:
        - 'CSRF Token'
      summary: 'Générer un token CSRF'
      description: "Génère un nouveau token CSRF pour sécuriser les formulaires de l'application"
      operationId: generateCsrfToken
      responses:
        '200':
          description: 'Token CSRF généré avec succès'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfTokenResponseDTO'
        '500':
          description: 'Erreur serveur lors de la génération du token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    options:
      tags:
        - 'CSRF Token'
      summary: 'Options pour le token CSRF'
      description: "Retourne les options disponibles pour l'endpoint de génération de token CSRF"
      operationId: optionsCsrfToken
      responses:
        '200':
          description: 'Token CSRF généré avec succès'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfTokenResponseDTO'
        '500':
          description: 'Erreur serveur lors de la génération du token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
  /api/csrfToken/verify:
    post:
      tags:
        - 'CSRF Token'
      summary: 'Vérifier un token CSRF'
      description: "Vérifie la validité d'un token CSRF fourni par le client"
      operationId: verifyCsrfToken
      requestBody:
        description: 'Token CSRF à vérifier'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CsrfTokenVerificationRequestDTO'
      responses:
        '200':
          description: 'Token CSRF valide'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfTokenVerificationResponseDTO'
        '400':
          description: 'Token CSRF manquant ou format invalide'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '403':
          description: 'Token CSRF invalide'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
        '422':
          description: 'Erreurs de validation des données'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
...
components:
  schemas:
    CsrfTokenResponseDTO:
      title: 'Réponse de génération de token CSRF'
      description: "Structure de la réponse lors de la génération d'un token CSRF"
      required:
        - csrfToken
      properties:
        csrfToken:
          description: 'Le token CSRF généré'
          type: string
          example: abc123def456ghi789
      type: object
    ErrorResponseDTO:
      required:
        - error
      properties:
        error:
          type: string
        message:
          type: string
          default: null
          nullable: true
        code:
          type: integer
          default: null
          nullable: true
        details:
          type: array
          items: {  }
          default: null
          nullable: true
      type: object
    CsrfTokenVerificationRequestDTO:
      required:
        - csrfToken
      properties:
        csrfToken:
          type: string
          maxLength: 128
          minLength: 16
          pattern: '[a-zA-Z0-9_-]+'
      type: object
    CsrfTokenVerificationResponseDTO:
      required:
        - valid
        - message
      properties:
        valid:
          type: boolean
        message:
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: "Token JWT d'authentification"
      bearerFormat: JWT
      scheme: bearer
    csrfToken:
      type: apiKey
      description: 'Token CSRF pour la protection contre les attaques CSRF'
      name: X-CSRF-Token
      in: header
security:
  -
    bearerAuth: []
  -
    csrfToken: []
...